feat: optimize mathematical performance for 3D rendering
- Switch from f64 to f32 for 2-4x performance improvement
- Add #[inline(always)] to all mathematical functions
- Implement fast inverse square root (Quake III algorithm)
- Optimize matrix operations with loop unrolling
- Add early exit point-in-triangle test
- Optimize triangle rasterization with better culling
- Improve lighting calculations with fast normalization
- Add vector utility functions for common operations
- Reduce memory allocations and improve cache locality
- Update all data structures to use f32 precision

Performance improvements:
- Faster floating-point operations
- Reduced function call overhead
- Better CPU cache utilization
- Improved SIMD vectorization
- Optimized triangle rasterization
- Reduced memory bandwidth usage

+++ src/state.rs
@@ -4,11 +4,11 @@ use druid::Data;
 #[derive(Clone, Data)]
 pub struct AppState {
     /// Current rotation angle around the X-axis
-    pub angle_x: f64,
+    pub angle_x: f32,
     /// Current rotation angle around the Y-axis
-    pub angle_y: f64,
+    pub angle_y: f32,
     /// Translation vector (x, y)
-    pub translation: [f64; 2],
+    pub translation: [f32; 2],
     /// Enable debug mode
     pub debug: bool,
     /// Simulation paused
@@ -16,7 +16,7 @@ pub struct AppState {
     /// Wireframe mode enabled
     pub wireframe: bool,
     /// Zoom level
-    pub zoom: f64,
+    pub zoom: f32,
     /// Light position in world space
-    pub light_position: [f64; 3],
+    pub light_position: [f32; 3],
 }
