+++ src/vertex.rs
@@ -1,6 +1,6 @@
 /// Vertex structure with position, screen position, and normal
 pub struct Vertex {
-    pub position: [f64; 3],
-    pub screen_position: [f64; 2],
-    pub normal: [f64; 3],
+    pub position: [f32; 3],
+    pub screen_position: [f32; 2],
+    pub normal: [f32; 3],
 }
